+++ python/GafferUITest/StandardNodeGadgetTest.py
@@ -308,5 +308,47 @@ class StandardNodeGadgetTest( GafferUITest.TestCase ) :
 		Gaffer.Metadata.deregisterValue( n, "nodeGadget:minWidth" )
 		assertMinWidth( g, 10.0 )
 
+	def testNoduleLabelVisibility( self ) :
+
+		n = Gaffer.Node()
+		n.addChild( Gaffer.FloatPlug( "fIn", direction = Gaffer.Plug.Direction.In ) )
+		n.addChild( Gaffer.FloatPlug( "fOut", direction = Gaffer.Plug.Direction.Out ) )
+		g = GafferUI.StandardNodeGadget( n )
+
+		fIn = g.nodule( n["fIn"] )
+		fOut = g.nodule( n["fOut"] )
+
+		self.assertFalse( fIn.getLabelVisible() )
+		self.assertFalse( fOut.getLabelVisible() )
+
+		Gaffer.Metadata.registerValue( Gaffer.Node, "nodeGadget:inputNoduleLabelsVisible", True )
+		self.assertTrue( fIn.getLabelVisible() )
+		self.assertFalse( fOut.getLabelVisible() )
+
+		Gaffer.Metadata.registerValue( Gaffer.Node, "nodeGadget:outputNoduleLabelsVisible", True )
+		self.assertTrue( fIn.getLabelVisible() )
+		self.assertTrue( fOut.getLabelVisible() )
+
+		n.addChild( Gaffer.IntPlug( "iIn", direction = Gaffer.Plug.Direction.In ) )
+		n.addChild( Gaffer.IntPlug( "iOut", direction = Gaffer.Plug.Direction.Out ) )
+
+		iIn = g.nodule( n["iIn"] )
+		iOut = g.nodule( n["iOut"] )
+
+		self.assertTrue( iIn.getLabelVisible() )
+		self.assertTrue( iOut.getLabelVisible() )
+
+		Gaffer.Metadata.registerValue( Gaffer.Node, "nodeGadget:inputNoduleLabelsVisible", False )
+		self.assertFalse( fIn.getLabelVisible() )
+		self.assertFalse( iIn.getLabelVisible() )
+		self.assertTrue( fOut.getLabelVisible() )
+		self.assertTrue( iOut.getLabelVisible() )
+
+		Gaffer.Metadata.registerValue( Gaffer.Node, "nodeGadget:outputNoduleLabelsVisible", False )
+		self.assertFalse( fIn.getLabelVisible() )
+		self.assertFalse( iIn.getLabelVisible() )
+		self.assertFalse( fOut.getLabelVisible() )
+		self.assertFalse( iOut.getLabelVisible() )
+
 if __name__ == "__main__":
 	unittest.main()